chore: refactor physics voxel world system
